Action: Add docker login variables
docker-entrypoint.sh can now handle variables that are used with docker
login. If a docker login username and password are provided, then
docker login will be ran with those input parameters.

A registry may be provided. If no registry is provided, then the login
will default to Docker Hub as that is the default login registry for
the Docker CLI.

Currently this only supports running docker login from the action
container. If the user wants to use the COPY_STACK_FILE option,
then logging into a registry won't matter since the login occurs
locally and not on the remote machine. If the user is using the
COPY_STACK_FILE option, then they will need to add the docker login
command manually to the INPUT_PRE_DEPLOYMENT_COMMAND_ARGS.